Rename formatter package to format to match the name of interface.
