+++ format/clock.go
@@ -1,4 +1,4 @@
-package formatter
+package format
 
 import (
 	"fmt"
