+++ gadget/timer/preset.go
@@ -2,15 +2,15 @@ package timer
 
 import (
 	"github.com/simp7/times"
-	"github.com/simp7/times/object/formatter"
+	"github.com/simp7/times/format"
 	"time"
 )
 
-//Standard returns gadget.Timer that implements object.Standard and formatter.Standard adopting minimum unit as second.
-//Hangul returns gadget.Timer that implements object.Standard and formatter.Hangul adopting minimum unit as second.
-//Detail returns gadget.Timer that implements object.Accurate and formatter.Detail adopting minimum unit as millisecond.
+//Standard returns gadget.Timer that implements object.Standard and format.Standard adopting minimum unit as second.
+//Hangul returns gadget.Timer that implements object.Standard and format.Hangul adopting minimum unit as second.
+//Detail returns gadget.Timer that implements object.Accurate and format.Detail adopting minimum unit as millisecond.
 var (
-	Standard = func(t times.Object) times.Gadget { return New(time.Second, formatter.Standard, t) }
-	Hangul   = func(t times.Object) times.Gadget { return New(time.Second, formatter.Hangul, t) }
-	Detail   = func(t times.Object) times.Gadget { return New(time.Millisecond, formatter.Detail, t) }
+	Standard = func(t times.Object) times.Gadget { return New(time.Second, format.Standard, t) }
+	Hangul   = func(t times.Object) times.Gadget { return New(time.Second, format.Hangul, t) }
+	Detail   = func(t times.Object) times.Gadget { return New(time.Millisecond, format.Detail, t) }
 )
